+++ src/pwg-param-info-private.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "pwg-param-info.h"
+
+G_BEGIN_DECLS
+
+PwgParamInfo *_pwg_param_info_new(guint id, guint flags);
+
+G_END_DECLS